feat(updated): add useSplash hook
