+++ src/screens/Splash/Splash.tsx
@@ -1,16 +1,9 @@
-import React, {useCallback, useEffect} from 'react';
+import React, {useEffect} from 'react';
 
 import {View} from 'react-native';
-import Animated, {
-  useAnimatedStyle,
-  useSharedValue,
-  withTiming,
-  interpolate,
-  Easing,
-  runOnJS,
-} from 'react-native-reanimated';
-import RNBootSplash from 'react-native-bootsplash';
+import Animated from 'react-native-reanimated';
 import {isIOS} from 'services/PlatformService';
+import {useSplash} from './hooks';
 import {styles} from './styles';
 
 interface IProps {
@@ -19,90 +12,19 @@ interface IProps {
 }
 
 export const Splash = ({onAnimationEnd, onFadeStart}: IProps) => {
-  const opacity = useSharedValue(1);
-  const animatedValue = useSharedValue(0);
-
-  const imageAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      transform: [
-        {translateX: interpolate(animatedValue.value, [0, 1], [0, -90])},
-      ],
-    };
-  });
-
-  const textAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      opacity: animatedValue.value,
-      transform: [
-        {scale: animatedValue.value},
-        {
-          translateX: interpolate(animatedValue.value, [0, 1], [0, 35]),
-        },
-      ],
-    };
-  });
-
-  const containerAnimatedStyle = useAnimatedStyle(() => {
-    return {
-      opacity: opacity.value,
-    };
-  });
-
-  const animateIOS = useCallback(() => {
-    setTimeout(() => {
-      RNBootSplash.hide().then(() => {
-        animatedValue.value = withTiming(
-          1,
-          {
-            duration: 300,
-            easing: Easing.out(Easing.ease),
-          },
-          () => {
-            if (onAnimationEnd) {
-              runOnJS(onAnimationEnd)();
-            }
-          },
-        );
-      });
-    }, 10);
-  }, [animatedValue, onAnimationEnd]);
-
-  const animateAndroid = useCallback(() => {
-    setTimeout(() => {
-      RNBootSplash.hide().then(() => {
-        animatedValue.value = withTiming(
-          1,
-          {
-            duration: 400,
-            easing: Easing.out(Easing.ease),
-          },
-          () => {
-            if (onFadeStart) {
-              runOnJS(onFadeStart)();
-            }
-            opacity.value = withTiming(
-              0,
-              {
-                duration: 300,
-                easing: Easing.out(Easing.ease),
-              },
-              () => {
-                if (onAnimationEnd) {
-                  runOnJS(onAnimationEnd)();
-                }
-              },
-            );
-          },
-        );
-      });
-    }, 300);
-  }, [animatedValue, onAnimationEnd, onFadeStart, opacity]);
+  const {
+    animateIOS,
+    animateAndroid,
+    containerAnimatedStyle,
+    imageAnimatedStyle,
+    textAnimatedStyle,
+  } = useSplash();
 
   useEffect(() => {
     if (isIOS) {
-      animateIOS();
+      animateIOS(onAnimationEnd);
     } else {
-      animateAndroid();
+      animateAndroid(onFadeStart, onAnimationEnd);
     }
   }, [animateAndroid, animateIOS]);
