+++ src/screens/Splash/hooks/index.tsx
@@ -1,0 +1,1 @@
+export {useSplash} from './useSplash';
